+++ app/src/main/java/io/omniedge/SingletonHolder.kt
@@ -8,6 +8,7 @@ package io.omniedge
  */
 open class SingletonHolder<out T : Any, in A>(creator: (A) -> T) {
     private var creator: ((A) -> T)? = creator
+
     @Volatile
     private var instance: T? = null