+++ app/src/main/java/io/omniedge/data/bean/Response.kt
@@ -1,10 +1,11 @@
 package io.omniedge.data.bean
 
+import com.google.gson.JsonObject
 import com.google.gson.annotations.SerializedName
 
 data class Response(
     val message: String?,
-    val errors: ErrorData?,
+    val errors: JsonObject?,
 )
 
 data class ErrorData(val password: String?)
@@ -19,22 +20,42 @@ data class LoginResponse(
 data class RegisterDeviceResponse(val message: String?, val data: DeviceData?)
 data class ListNetworkResponse(val message: String?, val data: List<NetworkData> = arrayListOf())
 data class JoinNetworkResponse(val message: String?, val data: JoinNetworkData)
+data class CreateNetworkResponse(val message: String?, val data: NetworkData)
 
 data class DeviceData(
-    val uuid: String,
+    val id: String,
     val name: String?,
-    val os: String?,
+    val platform: String?,
+    @SerializedName("hardware_id") val hardwareId: String?,
     @SerializedName("virtual_ip") val virtualIp: String?,
-    @SerializedName("last_seen") val lastSeen: String?,
+    @SerializedName("created_at") val created_at: String?,
     val online: Boolean = false,
 )
 
+data class UserData(
+    val id: String,
+    val name: String?,
+    val email: String?,
+    val picture: String?,
+    @SerializedName("last_login_ip") val lastLoginIp: String?,
+    @SerializedName("last_login_at") val lastLoginAt: String?,
+    val status: Int,
+    @SerializedName("created_at") val createdAt: String?,
+    @SerializedName("updated_at") val updatedAt: String?,
+    val role: Int,
+    @SerializedName("joined_at") val joinedAt: String?,
+)
+
 data class NetworkData(
-    val uuid: String,
+    val id: String,
     val name: String?,
     @SerializedName("ip_range") val ipRange: String?,
-    val role: String?,
+    val role: Int,
+    val usersCount: Int,
+    val devicesCount: Int,
     val devices: List<DeviceData>?,
+    val users: List<UserData>?,
+    val server: ServerData,
 )
 
 data class JoinNetworkData(
@@ -43,6 +64,6 @@ data class JoinNetworkData(
     @SerializedName("virtual_ip") val virtualIp: String,
     @SerializedName("subnet_mask") val subnetMask: String,
     @SerializedName("server") val server: ServerData,
-) {
-    data class ServerData(val name: String, val country: String, val host: String)
-}
+)
+
+data class ServerData(val id: String, val name: String, val country: String, val host: String)