+++ app/src/main/java/io/omniedge/ui/fragment/BaseFragment.kt
@@ -6,10 +6,11 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
+import android.widget.TextView
 import androidx.annotation.LayoutRes
 import androidx.annotation.StringRes
 import androidx.fragment.app.Fragment
+import com.google.android.material.snackbar.Snackbar
 import io.omniedge.PageView
 import io.reactivex.disposables.CompositeDisposable
 
@@ -21,21 +22,41 @@ abstract class BaseFragment : Fragment(), PageView {
 
     private val compositeDisposable = CompositeDisposable()
 
-    fun toast(@StringRes msg: Int) {
+    fun toast(view: View, @StringRes msg: Int) {
         activity?.runOnUiThread {
-            Toast.makeText(context, msg, Toast.LENGTH_SHORT).show()
+            Snackbar.make(view, msg, Snackbar.LENGTH_SHORT)
+                .apply {
+                    this.view
+                        .findViewById<TextView>(com.google.android.material.R.id.snackbar_text)
+                        .maxLines = 5
+                }
+                .show()
         }
     }
 
-    private fun toast(msg: String) {
+    private fun toast(view: View, msg: String) {
         activity?.runOnUiThread {
-            Toast.makeText(context, msg, Toast.LENGTH_SHORT).show()
+            Snackbar.make(view, msg, Snackbar.LENGTH_SHORT)
+                .apply {
+                    this.view
+                        .findViewById<TextView>(com.google.android.material.R.id.snackbar_text)
+                        .maxLines = 5
+                }
+                .show()
         }
     }
 
     @LayoutRes
-    open fun getLayoutRes(): Int {
+    protected open fun getLayoutRes(): Int {
         return -1
+    }
+
+    protected open fun getLayoutView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return null
     }
 
     @SuppressLint("ResourceType")
@@ -49,11 +70,21 @@ abstract class BaseFragment : Fragment(), PageView {
             return inflater.inflate(getLayoutRes(), container, false)
         }
 
+        val layoutRes = getLayoutRes()
+        if (layoutRes > 0) {
+            return inflater.inflate(layoutRes, container, false)
+        } else {
+            val layoutView = getLayoutView(inflater, container, savedInstanceState)
+            if (layoutView != null) {
+                return layoutView
+            }
+        }
+
         return super.onCreateView(inflater, container, savedInstanceState)
     }
 
     override fun getCurContext(): Context {
-        return context!!
+        return requireContext()
     }
 
     override fun compositeDisposable(): CompositeDisposable {
@@ -61,6 +92,6 @@ abstract class BaseFragment : Fragment(), PageView {
     }
 
     override fun showToast(msg: String) {
-        toast(msg)
+        toast(requireActivity().findViewById(android.R.id.content), msg)
     }
 }