+++ CHANGES.md
@@ -1,4 +1,7 @@
 # Change log for OmniEdge Android
+## Version 0.2.1
+* Migrated to new backend APIs.
+* Add network creation logic on account without networks.
 
 ## Version 0.2.0